+++ src/TabRow/index.ts
@@ -1,0 +1,8 @@
+import TabRow from './tab_row';
+import {
+	TabRowItem as TabRowItemTemp,
+} from './tab_row';
+
+
+export default TabRow;
+export type TabRowItem = TabRowItemTemp;