+++ src/utility/style_utility.tsx
@@ -85,7 +85,7 @@ export function getEmphasisClasses(emphasis: EmphasisType, color: ColorOptionTyp
         
         // Setting the hover effect
         if (options.hoverable) {
-            if (emphasis.includes("background") && !options.transparentBackground) {
+            if (emphasis.includes("background")) {
                 c += ` hover-background--${color}-light`;
             } else {
                 c += ` hover-border--${color}-${emphasis === 'none' ? 'light' : 'normal'}`;